requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
executing==1.2.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.3
wcwidth==0.2.6
